Engine: Added ShakeScreen() special action
Also implemented F12 to return to the title screen and added a move frequency variable to events that decides how often they move.

+++ src/com/fourisland/fourpuzzle/util/Interval.java
@@ -22,12 +22,12 @@ public class Interval {
     
     public static Interval createTickInterval(float ticks)
     {
-        return createMillisInterval((int) (Game.FPS * ticks));
+        return createMillisInterval(Game.FPS * ticks);
     }
     
-    public static Interval createMillisInterval(int millis)
+    public static Interval createMillisInterval(float millis)
     {
-        return new Interval(millis*1000000);
+        return new Interval((int) (millis*1000000));
     }
     
     private long last = System.nanoTime();
